// src/app/models/interfaces/win32-bios.interface.ts
export interface IWin32Bios {
  BuildNumber: string;
  CodeSet: string;
  CurrentLanguage: string;
  Description: string;
  EmbeddedControllerMajorVersion: string;
  EmbeddedControllerMinorVersion: string;
  InstallDate: string;
  Manufacturer: string;
  Name: string;
  PrimaryBIOS: string;
  SerialNumber: string;
  SoftwareElementID: string;
  SystemBiosMajorVersion: string;
  SystemBiosMinorVersion: string;
  TargetOperatingSystem: string;
  Version: string;
}