// src/app/models/interfaces/win32-baseboard.interface.ts
export interface IWin32Baseboard {
  Depth: string;
  Description: string;
  Height: string;
  HostingBoard: string;
  InstallDate: string;
  Manufacturer: string;
  Model: string;
  Name: string;
  PartNumber: string;
  PoweredOn: string;
  Product: string;
  Removable: string;
  Replaceable: string;
  RequirementsDescription: string;
  RequiresDaughterBoard: string;
  SerialNumber: string;
  SpecialRequirements: string;
  Status: string;
  Weight: string;
  Width: string;
}